backend/requirements.txt
fastapi
python-multipart
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary
redis
langgraph
langchain
langchain-community
groq
ollama
chromadb
playwright
python-dotenv
httpx
pydantic
email-validator
pypdf
python-docx
openpyxl
aiosmtplib
imap-tools
beautifulsoup4
